feat : JWT token and password decorder has been intitalized in the project

+++ src/main/java/com/example/backend/Service/CustomerService.java
@@ -11,9 +11,9 @@ import java.util.Optional;
 public interface CustomerService {
     Customer saveCustomer(CustomerPojo customerPojo);
     List<Customer> getAllCustomers();
-    Customer getCustomerById(int id);
-    void deleteCustomerById(int id);
+    Customer getCustomerById(Long id);
+    void deleteCustomerById(Long id);
     Customer updateCustomer(CustomerPojo customerPojo);
-    Optional<Customer> findById(int id) ;
+    Optional<Customer> findById(Long id) ;
 
 }
